for option m adds h,4,5,6,7,8,9 rename r,w,x to 1,2,3; for option n remove s and adds; adds option d; remove option a,c,r; for option l adds r

+++ src/list.h
@@ -11,7 +11,7 @@ typedef struct node
 
 typedef struct linklist
 {
-    int count; // number of elements in LIST
+    long int count; // number of elements in LIST
     node first;
     node last;
 } * linklist;
